update HTML and CSS file,
the logo and header link text already fix,
below the header the main text and img fix the right margin and padding,

i made a mistake during the git commit i use the ammend and i dint know if all the update i done overwrite the last commit, this commit is just a div and css text change. the last commit supposed to be the big change.